caws_experiments/benchmarks/requirements.txt
Pillow==9.0.0
numpy==1.20
torch==1.10.0+cu113 torchvision==0.11.1+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
torchvision==0.9.0
python-igraph==0.10.6
squiggle==0.3.1